First Worthy Upload "SSH2 TeraTerm Session Manager"  SSHTTSM
SSHTTSM will try to use a tterm.ttl macro file loacted in the appdata... teraterm5 folder if it exists.  This release was devloped and tested using TeraTerm 5.5 x64.  Good luck with it.  Written using PureBasic ver 5.40  and as the sole creator I can certify it to be clean here on GitHub.  Not distributed anywhere else.