Username & Passwords enabled!!!
Happy to announce I have added ability to perform auto-login.  Username and password are Optional fields when saving or editing a session.  The [USR/PWD OK] item MUST be enabled (checked) to save credentials for the session and To enable auto-Login.  If that box is not checked when you click [Connect] then you will be prompted for credentials to complete your login.